A metabolite data set from Arabidopsis leaves by GC-TOF/MS

A metabolite data set. The Arabidopsis metabolome of the aerial regions of individual WT plants and the mto1 and tt4 mutants were analyzed by GC-TOF/MS.

Details

50 samples (WT, n = 17; mto1, n = 13; and tt4, n = 20, biological replicates).

A matrix containing 59 metabolites (rows) and 50 observations (columns).

MetaboLights accession no.: MTBLS40

For comparisons with data from roots (Fukushima et al. 2011) we selected 59 commonly-detected metabolites in both datasets using MetMask http://metmask.sourceforge.net.

References

Miyako Kusano, Atsushi Fukushima et al. BMC Syst Biol 2007 1:53