Additional distance functions correlation distance (1-r)

Additional distance functions Correlation distance (1-r)

cor.dist(data, methods = "pearson", absolute = FALSE)

Arguments

data
a data matrix ([data.frame object] row: metabolites, col: samples or replicates)

methods
a character string indicating which correlation coefficient is to be calculated. One of "pearson" (default), "spearman", or "kendall" can be abbreviated.

absolute
TRUE means that absolute value of the correlation coefficient is used (Default: FALSE).

Value

the resulting correlation matrix

Details

These functions were originally from 'hybridHclust' package. We modified the functions slightly. See also the reference manual in detail.

Examples

cor.dist(as.matrix(t(iris[,1:4])))

#>              Sepal.Length Sepal.Width Petal.Length Petal.Width
#> Sepal.Length    0.0000000    1.117570   0.12824622  0.18205887
#> Sepal.Width     1.1175698    0.000000   1.42844010  1.36612593
#> Petal.Length    0.1282462    1.428440   0.00000000  0.03713457
#> Petal.Width     0.1820589    1.366126   0.03713457  0.00000000